[fix] change type of role_id

+++ app/src/main/java/dev/abelab/crs/db/entity/User.java
@@ -54,7 +54,7 @@ public class User {
      *
      * @mbg.generated
      */
-    private Byte roleId;
+    private Integer roleId;
 
     /**
      *
